+++ Not.cs
@@ -4,8 +4,8 @@ public class Not : IInstruction{
 	public Not(){}
 	public uint Val{
 		get{
-			uint retVal = 49 << 24; //0011 0001
-			return retVal;
+			int retVal = 49 << 24; //0011 0001
+			return (uint)retVal;
 		}
 	}
 }
