+++ Print.cs
@@ -4,8 +4,8 @@ public class Print : IInstruction{
 	public Print(){}
 	public uint Val{
 		get{
-			uint retVal = 13 << 28; //1101
-			return retVal;
+			int retVal = 13 << 28; //1101
+			return (uint)retVal;
 		}
 	}
 }
