+++ Push.cs
@@ -1,16 +1,16 @@
 using System;
 
 public class Push : IInstruction{
-	uint val;
+	int val;
 
-	public Push(uint input){
+	public Push(int input){
 		val = input;
 	}
 	public uint Val{
 		get{
-			uint retVal = 15 << 28; //1111
+			int retVal = 15 << 28; //1111
 			retVal = retVal | val; //assumes top 4 bits of Val are 0000
-			return retVal;
+			return (uint)retVal;
 		}
 	}
 }
